Fixed outdated import statement.
The Problem
An outdated import statements was causing an error in emojis.py.

Due the the try-catch phrase on lines 28-37 of utils.py a misleading error message
suggests that the markdown verison is out of date.

The Solution
The error is in fact caused by an outdated import statement.
See lines 14 and 15 of the updated django source code file at the link below.

https://github.com/django/django/blob/aa5fd84f53f09338d01a3cfd9fa6ab08e418fe00/django/contrib/staticfiles/templatetags/staticfiles.py