Adjust regex so that it does not strip links inadvertently
The issue with the current implementation seems to stem from the
regex pattern mistakenly capturing and modifying links that should
be left untouched, especially when they contain colons (:) or
semicolons (;) that are part of a normal URL or anchor reference (#).

This was leading to issues where text would be lost where a paragraph
would contain more than one link and a semi-colon.

eg:

- for this input:
  - Citation 4. [footnote 4](#ftnt4) Citation 5. [footnote 5](#ftnt5) Citation 6; [footnote 6](#ftnt6)
- we were getting this output:
  - <p>Citation 4. <a href="#ftnt4">footnote 4</a> Citation 5. <a href=";">footnote 5</a></p>
- when what we wanted was this:
  - <p>Citation 4. <a href="#ftnt4">footnote 4</a> Citation 5. <a href="#ftnt5">footnote 5</a> Citation 6; <a href="#ftnt6">footnote 6</a></p>

This PR fixes this so that regular links are preserved, but keeps in
place the logic stripping bad protocols.

Closes this issue:
- https://github.com/agusmakmun/django-markdown-editor/issues/229